Rename duke package to dietbook
