+++ build.gradle
@@ -31,7 +31,7 @@ test {
 }
 
 application {
-    mainClassName = "seedu.duke.DietBook"
+    mainClassName = "seedu.dietbook.DietBook"
 }
 
 shadowJar {
