+++ src/main/java/seedu/dietbook/database/Store.java
@@ -1,6 +1,6 @@
-package seedu.duke.database;
+package seedu.dietbook.database;
 
-import seedu.duke.food.Food;
+import seedu.dietbook.food.Food;
 
 import java.util.ArrayList;
 import java.util.Arrays;
