+++ src/main/java/seedu/dietbook/food/Food.java
@@ -1,4 +1,4 @@
-package seedu.duke.food;
+package seedu.dietbook.food;
 
 
 /**
