+++ src/main/java/seedu/dietbook/list/FoodNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.duke.list;
+package seedu.dietbook.list;
 
 public class FoodNotFoundException extends Exception{
 }
