+++ src/test/java/seedu/dietbook/database/DataBaseTest.java
@@ -1,4 +1,4 @@
-package seedu.duke.database;
+package seedu.dietbook.database;
 
 import java.util.NoSuchElementException;
